Recreated UserRequest, dto for registering users
That's because User needs to be validated before registering. The
actual password must be 8-30 chars long, but in the database it has to be
encrypted and this form will be longer than 30 chars.

That's why I want to have a class doing all the validations for checking
if the user fields are valid, and another one representing an entity in
db.