Implemented activity status for users
User entity now have lastActive field which is LocalDateTime. It's value
says what was the last time user was connected to WebSocket, if it's
null it means that user is active now.

Activity status is changed based on connection and disconnection via
WebSocket.

Information about activity is also placed in ChatPreview dto class.

So it's initially send as a part of cha object with a certain user.
To have it updated, the frontend sends request to UserApi with list
of all loaded chats and receives a map (js object) of pairs - userId
with user's status.