Made Security Classes more configurable
To deploy on the AWS EC2, I need to make it possible to be able to set
the frontend origin in Security classes. I set it from a
aws.frontend.origin property.
I also created default config for the new local profile, which will be
used on my local machine, on aws I will provide config through command
line arguments