Fixed registration + implemented AWS Email Service
Included not activated users in terms of username and email being
unique in Registration process.
Created AWSEmailService which uses ESE, but it is only for demo
purposes, because my AWS is in ESE sandbox. Because of that, I cannot
send emails to unauthorized addresses, which basically makes it
impossible to make registration process with confirmation through
email. That's why I also made MockEmailService.